+++ CHANGELOG.md
@@ -1,6 +1,63 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [1.0.1-dev.1748611088] - 2025-05-30
+
+### 📝 Other Changes
+- style: improve responsiveness of movie list
+
+
+## [1.0.1-dev.1748610462] - 2025-05-30
+
+### 📝 Other Changes
+- style: fix badge hovering for movie card
+
+
+## [1.0.1-dev.1748610180] - 2025-05-30
+
+### 📝 Other Changes
+- style: improve responsiveness of discover tab
+
+
+
+## [1.1.0-dev.1748607298] - 2025-05-30
+
+### 🚀 Features
+- feat: filter the movies you watched based on ratings
+- feat: complete local semantic release system setup
+- feat: add test feature for version bump demo
+
+### 🐛 Bug Fixes
+- fix: fix watchlist's problem when filtering on media
+- fix: resolve semantic release promotion workflow and clean version numbers
+
+### 📝 Other Changes
+- style: improve responsiveness and layout in watched movies tab
+
+
+## [1.1.0-dev.1748606113] - 2025-05-30
+
+### 🚀 Features
+- feat: filter the movies you watched based on ratings
+- feat: complete local semantic release system setup
+- feat: add test feature for version bump demo
+
+### 🐛 Bug Fixes
+- fix: fix watchlist's problem when filtering on media
+- fix: resolve semantic release promotion workflow and clean version numbers
+
+
+## [1.1.0-dev.1748605329] - 2025-05-30
+
+### 🚀 Features
+- feat: complete local semantic release system setup
+- feat: add test feature for version bump demo
+
+### 🐛 Bug Fixes
+- fix: fix watchlist's problem when filtering on media
+- fix: resolve semantic release promotion workflow and clean version numbers
+
+
 ## [1.1.0-dev.1748602143] - 2025-05-30
 
 ### 🚀 Features